requirements.txt
browser-use
fastapi==0.104.1
numpy
uvicorn==0.24.0
pydantic==2.10.1
pydantic-settings==2.0.3
websockets==12.0
asyncio==3.4.3
python-multipart==0.0.6
aiofiles==23.2.1
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.3
anthropic>=0.19.1
langchain>=0.1.0
langchain_anthropic>=0.1.1
openai>=1.10.0
langchain_openai>=0.0.5
google-generativeai>=0.3.1
langchain_google_genai>=0.1.1